requirements.txt
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
schedule>=1.2.0
gspread>=5.12.0
google-auth>=2.23.0
